requirements.txt
fastapi
uvicorn
streamlit
sentence-transformers
openai
faiss-cpu